Current motion detect code on RPI
This is the current motion detection software that is programmed and successful on the RPI. It contains the software to engage the ultrasonic sensor, detect motion, return the distance of the object, take an image, and the solenoid_active() function. Everything is successful in testing except the solenoid_active() function which is unsuccessful.